add: completion item

+++ app/Http/Controllers/ChecklistController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Checklist;
 use App\ChecklistItem;
 use App\Http\Resources\ChecklistCollection;
+use App\Http\Services\Completion;
 use Auth;
+use Carbon\Carbon;
 use Illuminate\Http\Request;
 
 class ChecklistController extends Controller
@@ -53,9 +55,11 @@ class ChecklistController extends Controller
                 'links' => $links,
                 'data'  => $data,
             );
+            
+            return response()->json($result);
         }
         
-        return response()->json($result);
+        return response()->json(['status' => 404, 'error' => 'Not Found'], 404);
     }
     
     /**
@@ -67,9 +71,11 @@ class ChecklistController extends Controller
      */
     public function show($id, Checklist $model)
     {
-        $result = $model->where('id', $id)->first();
+        if ($result = $model->where('id', $id)->first()) {
+            return new ChecklistCollection($result);
+        }
         
-        return new ChecklistCollection($result);
+        return response()->json(['status' => 404, 'error' => 'Not Found'], 404);
     }
     
     /**
@@ -84,6 +90,7 @@ class ChecklistController extends Controller
     {
         $this->validate($request, [
             'data'                          => 'required',
+            'data.attributes'               => 'required',
             'data.attributes.object_domain' => 'required',
             'data.attributes.object_id'     => 'required',
             'data.attributes.description'   => 'required',
@@ -102,7 +109,7 @@ class ChecklistController extends Controller
                 $ModelItem->Create($data_item);
             }
             
-            return response()->json(['status' => 'success']);
+            return new ChecklistCollection($c);
         }
         
         return response()->json(['status' => 'fail'], 500);
@@ -117,16 +124,18 @@ class ChecklistController extends Controller
      */
     public function delete($id, Checklist $model)
     {
-        if ($model->destroy($id)) {
-            return response()->json(['status' => 'success']);
+        if ($d = $model->find($id)) {
+            if ($model->destroy($id)) {
+                return response()->json(['status' => '204']);
+            }
+            
+            return response()->json(['status' => 500, 'error' => 'Server Error'], 500);
         }
         
-        return response()->json(['status' => 'fail'], 500);
+        return response()->json(['status' => 404, 'error' => 'Not Found'], 404);
     }
     
     /**
-     * Remove the specified resource from storage.
-     *
      * @param  int      $id
      * @param Request   $request
      * @param Checklist $model
@@ -138,6 +147,7 @@ class ChecklistController extends Controller
             'data'                          => 'required',
             'data.type'                     => 'required',
             'data.id'                       => 'required',
+            'data.attributes'               => 'required',
             'data.attributes.object_domain' => 'required',
             'data.attributes.object_id'     => 'required',
             'data.attributes.description'   => 'required',
@@ -145,14 +155,84 @@ class ChecklistController extends Controller
             'data.links.self'               => 'required',
         ]);
         
-        if ($d = $model->find($id)) {
-            if ($d->fill($request->input('data.attributes'))->save()) {
-                return response()->json(['status' => '200']);
+        if ($c = $model->find($id)) {
+            $data               = $request->input('data.attributes');
+            $data['updated_by'] = current_user()->id;
+            if ($c->fill($data)->save()) {
+                return new ChecklistCollection($c);
             }
             
             return response()->json(['status' => 500, 'error' => 'Server Error'], 500);
         }
         
         return response()->json(['status' => 404, 'error' => 'Not Found'], 404);
+    }
+    
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\Response
+     */
+    public function complete(Request $request)
+    {
+        $this->validate($request, [
+            'data' => 'required',
+        ]);
+        
+        $result        = [];
+        $checklist_ids = [];
+        
+        foreach ($request->input('data') as $id) {
+            if ($d = ChecklistItem::find($id)) {
+                $d = $d->first();
+                
+                $d->is_completed = true;
+                $d->completed_at = Carbon::now();
+                $d->save();
+                
+                $result[]        = $d;
+                $checklist_ids[] = $d->checklist_id;
+            }
+        }
+        
+        foreach ($checklist_ids as $checklist_id) {
+            //TODO : use event listener better
+            Completion::ChecklistCompletionChecker($checklist_id);
+        }
+        
+        return response()->json(['data' => $result], 200);
+    }
+    
+    /**
+     * @param Request $request
+     * @return \Illuminate\Http\Response
+     */
+    public function incomplete(Request $request)
+    {
+        $this->validate($request, [
+            'data' => 'required',
+        ]);
+        
+        $result        = [];
+        $checklist_ids = [];
+        
+        foreach ($request->input('data') as $id) {
+            if ($d = ChecklistItem::find($id)) {
+                $d = $d->first();
+                
+                $d->is_completed = false;
+                $d->completed_at = null;
+                $d->save();
+                
+                $result[]        = $d;
+                $checklist_ids[] = $d->checklist_id;
+            }
+        }
+        
+        foreach ($checklist_ids as $checklist_id) {
+            //TODO : use event listener better
+            Completion::ChecklistCompletionChecker($checklist_id);
+        }
+        
+        return response()->json(['data' => $result], 200);
     }
 }
